+++ crates/nu-cmd-extra/src/extra/mod.rs
@@ -88,7 +88,14 @@ pub fn add_extra_command_context(mut engine_state: EngineState) -> EngineState {
             strings::format::FileSize,
             strings::format::FormatDuration,
             strings::encode_decode::EncodeHex,
-            strings::encode_decode::DecodeHex
+            strings::encode_decode::DecodeHex,
+            strings::str_::case::Str,
+            strings::str_::case::StrCamelCase,
+            strings::str_::case::StrKebabCase,
+            strings::str_::case::StrPascalCase,
+            strings::str_::case::StrScreamingSnakeCase,
+            strings::str_::case::StrSnakeCase,
+            strings::str_::case::StrTitleCase
         );
 
         bind_command!(formats::ToHtml, formats::FromUrl);
