+++ crates/nu-cmd-extra/src/extra/strings/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod encode_decode;
 pub(crate) mod format;
+pub(crate) mod str_;
