+++ crates/nu-cmd-extra/src/extra/strings/str_/case/camel_case.rs
@@ -1,21 +1,22 @@
-use inflector::cases::screamingsnakecase::to_screaming_snake_case;
+use inflector::cases::camelcase::to_camel_case;
 use nu_protocol::ast::Call;
 use nu_protocol::engine::{Command, EngineState, Stack};
 use nu_protocol::{
     Category, Example, PipelineData, ShellError, Signature, Span, SyntaxShape, Type, Value,
 };
 
-use crate::operate;
+use super::operate;
+
 #[derive(Clone)]
 pub struct SubCommand;
 
 impl Command for SubCommand {
     fn name(&self) -> &str {
-        "str screaming-snake-case"
+        "str camel-case"
     }
 
     fn signature(&self) -> Signature {
-        Signature::build("str screaming-snake-case")
+        Signature::build("str camel-case")
             .input_output_types(vec![
                 (Type::String, Type::String),
                 (
@@ -35,11 +36,11 @@ impl Command for SubCommand {
     }
 
     fn usage(&self) -> &str {
-        "Convert a string to SCREAMING_SNAKE_CASE."
+        "Convert a string to camelCase."
     }
 
     fn search_terms(&self) -> Vec<&str> {
-        vec!["convert", "style", "underscore", "convention"]
+        vec!["convert", "style", "caps", "convention"]
     }
 
     fn run(
@@ -49,34 +50,34 @@ impl Command for SubCommand {
         call: &Call,
         input: PipelineData,
     ) -> Result<PipelineData, ShellError> {
-        operate(engine_state, stack, call, input, &to_screaming_snake_case)
+        operate(engine_state, stack, call, input, &to_camel_case)
     }
 
     fn examples(&self) -> Vec<Example> {
         vec![
             Example {
-                description: "convert a string to SCREAMING_SNAKE_CASE",
-                example: r#" "NuShell" | str screaming-snake-case"#,
-                result: Some(Value::test_string("NU_SHELL")),
+                description: "convert a string to camelCase",
+                example: " 'NuShell' | str camel-case",
+                result: Some(Value::test_string("nuShell")),
             },
             Example {
-                description: "convert a string to SCREAMING_SNAKE_CASE",
-                example: r#" "this_is_the_second_case" | str screaming-snake-case"#,
-                result: Some(Value::test_string("THIS_IS_THE_SECOND_CASE")),
+                description: "convert a string to camelCase",
+                example: "'this-is-the-first-case' | str camel-case",
+                result: Some(Value::test_string("thisIsTheFirstCase")),
             },
             Example {
-                description: "convert a string to SCREAMING_SNAKE_CASE",
-                example: r#""this-is-the-first-case" | str screaming-snake-case"#,
-                result: Some(Value::test_string("THIS_IS_THE_FIRST_CASE")),
+                description: "convert a string to camelCase",
+                example: " 'this_is_the_second_case' | str camel-case",
+                result: Some(Value::test_string("thisIsTheSecondCase")),
             },
             Example {
-                description: "convert a column from a table to SCREAMING_SNAKE_CASE",
-                example: r#"[[lang, gems]; [nu_test, 100]] | str screaming-snake-case lang"#,
+                description: "convert a column from a table to camelCase",
+                example: r#"[[lang, gems]; [nu_test, 100]] | str camel-case lang"#,
                 result: Some(Value::List {
                     vals: vec![Value::Record {
                         span: Span::test_data(),
                         cols: vec!["lang".to_string(), "gems".to_string()],
-                        vals: vec![Value::test_string("NU_TEST"), Value::test_int(100)],
+                        vals: vec![Value::test_string("nuTest"), Value::test_int(100)],
                     }],
                     span: Span::test_data(),
                 }),
