+++ crates/nu-cmd-extra/src/extra/strings/str_/case/kebab_case.rs
@@ -5,7 +5,7 @@ use nu_protocol::{
     Category, Example, PipelineData, ShellError, Signature, Span, SyntaxShape, Type, Value,
 };
 
-use crate::operate;
+use super::operate;
 
 #[derive(Clone)]
 pub struct SubCommand;
