+++ crates/nu-cmd-extra/src/extra/strings/str_/case/title_case.rs
@@ -1,30 +1,30 @@
-use inflector::cases::pascalcase::to_pascal_case;
+use inflector::cases::titlecase::to_title_case;
 use nu_protocol::ast::Call;
 use nu_protocol::engine::{Command, EngineState, Stack};
 use nu_protocol::{
     Category, Example, PipelineData, ShellError, Signature, Span, SyntaxShape, Type, Value,
 };
 
-use crate::operate;
+use super::operate;
 
 #[derive(Clone)]
 pub struct SubCommand;
 
 impl Command for SubCommand {
     fn name(&self) -> &str {
-        "str pascal-case"
+        "str title-case"
     }
 
     fn signature(&self) -> Signature {
-        Signature::build("str pascal-case")
+        Signature::build("str title-case")
             .input_output_types(vec![
                 (Type::String, Type::String),
-                (Type::Table(vec![]), Type::Table(vec![])),
-                (Type::Record(vec![]), Type::Record(vec![])),
                 (
                     Type::List(Box::new(Type::String)),
                     Type::List(Box::new(Type::String)),
                 ),
+                (Type::Table(vec![]), Type::Table(vec![])),
+                (Type::Record(vec![]), Type::Record(vec![])),
             ])
             .allow_variants_without_examples(true)
             .rest(
@@ -36,11 +36,11 @@ impl Command for SubCommand {
     }
 
     fn usage(&self) -> &str {
-        "Convert a string to PascalCase."
+        "Convert a string to Title Case."
     }
 
     fn search_terms(&self) -> Vec<&str> {
-        vec!["convert", "style", "caps", "upper", "convention"]
+        vec!["convert", "style", "convention"]
     }
 
     fn run(
@@ -50,34 +50,29 @@ impl Command for SubCommand {
         call: &Call,
         input: PipelineData,
     ) -> Result<PipelineData, ShellError> {
-        operate(engine_state, stack, call, input, &to_pascal_case)
+        operate(engine_state, stack, call, input, &to_title_case)
     }
 
     fn examples(&self) -> Vec<Example> {
         vec![
             Example {
-                description: "convert a string to PascalCase",
-                example: "'nu-shell' | str pascal-case",
-                result: Some(Value::test_string("NuShell")),
+                description: "convert a string to Title Case",
+                example: "'nu-shell' | str title-case",
+                result: Some(Value::test_string("Nu Shell")),
             },
             Example {
-                description: "convert a string to PascalCase",
-                example: "'this-is-the-first-case' | str pascal-case",
-                result: Some(Value::test_string("ThisIsTheFirstCase")),
+                description: "convert a string to Title Case",
+                example: "'this is a test case' | str title-case",
+                result: Some(Value::test_string("This Is A Test Case")),
             },
             Example {
-                description: "convert a string to PascalCase",
-                example: "'this_is_the_second_case' | str pascal-case",
-                result: Some(Value::test_string("ThisIsTheSecondCase")),
-            },
-            Example {
-                description: "convert a column from a table to PascalCase",
-                example: r#"[[lang, gems]; [nu_test, 100]] | str pascal-case lang"#,
+                description: "convert a column from a table to Title Case",
+                example: r#"[[title, count]; ['nu test', 100]] | str title-case title"#,
                 result: Some(Value::List {
                     vals: vec![Value::Record {
                         span: Span::test_data(),
-                        cols: vec!["lang".to_string(), "gems".to_string()],
-                        vals: vec![Value::test_string("NuTest"), Value::test_int(100)],
+                        cols: vec!["title".to_string(), "count".to_string()],
+                        vals: vec![Value::test_string("Nu Test"), Value::test_int(100)],
                     }],
                     span: Span::test_data(),
                 }),
