+++ crates/nu-cmd-extra/src/extra/strings/str_/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod case;
+
+pub use case::*;
