+++ crates/nu-command/src/default_context.rs
@@ -177,7 +177,6 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
             SplitRow,
             SplitWords,
             Str,
-            StrCamelCase,
             StrCapitalize,
             StrContains,
             StrDistance,
@@ -187,16 +186,11 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
             StrJoin,
             StrReplace,
             StrIndexOf,
-            StrKebabCase,
             StrLength,
-            StrPascalCase,
             StrReverse,
-            StrScreamingSnakeCase,
-            StrSnakeCase,
             StrStartsWith,
             StrSubstring,
             StrTrim,
-            StrTitleCase,
             StrUpcase,
             FormatDate
         };
